Move internal-event logic into its own class
