+++ wp-cron-control-revisited/wp-cron-control-revisited.php
@@ -34,14 +34,11 @@ class Main {
 	private $namespace = 'wp-cron-control-revisited/v1';
 	private $secret    = null;
 
-	private $job_queue_size                  = 10;
-	private $job_queue_window_in_seconds     = 60;
-	private $job_execution_buffer_in_seconds = 15;
-	private $job_timeout_in_minutes          = 10;
-	private $job_concurrency_limit           = 10;
-
-	private $internal_jobs           = array();
-	private $internal_jobs_schedules = array();
+	public $job_queue_size                  = 10;
+	public $job_queue_window_in_seconds     = 60;
+	public $job_execution_buffer_in_seconds = 15;
+	public $job_timeout_in_minutes          = 10;
+	public $job_concurrency_limit           = 10;
 
 	private $cache_key_lock           = 'wpccr_lock';
 	private $cache_key_lock_timestamp = 'wpccr_lock_ts';
@@ -52,6 +49,8 @@ class Main {
 	private function __construct() {
 		// Load dependencies
 		require __DIR__ . '/includes/class-cron-options-cpt.php';
+		require __DIR__ . '/includes/class-internal-events.php';
+		require __DIR__ . '/includes/functions-internal-events.php';
 
 		// For now, leave WP-CLI alone
 		if ( defined( 'WP_CLI' ) && WP_CLI ) {
@@ -72,10 +71,6 @@ class Main {
 			// Core plugin functionality
 			$this->prepare();
 			add_action( 'rest_api_init', array( $this, 'rest_api_init' ) );
-			add_filter( 'cron_schedules', array( $this, 'register_internal_events_schedules' ) );
-			add_action( 'wp_loaded', array( $this, 'schedule_internal_events' ) );
-			add_action( 'wpccrij_force_publish_missed_schedules', array( $this, 'force_publish_missed_schedules' ) );
-			add_action( 'wpccrij_confirm_scheduled_posts', array( $this, 'confirm_scheduled_posts' ) );
 		} else {
 			add_action( 'admin_notices', array( $this, 'admin_notice' ) );
 		}
@@ -87,29 +82,6 @@ class Main {
 	private function prepare() {
 		// Authentication
 		$this->secret = WP_CRON_CONTROL_SECRET;
-
-		// Internal jobs
-		$this->internal_jobs = array(
-			array(
-				'schedule' => 'wpccrij_minute',
-				'action'   => 'wpccrij_force_publish_missed_schedules',
-			),
-			array(
-				'schedule' => 'wpccrij_ten_minutes',
-				'action'   => 'wpccrij_confirm_scheduled_posts',
-			),
-		);
-
-		$this->internal_jobs_schedules = array(
-			'wpccrij_minute' => array(
-				'interval' => 1 * MINUTE_IN_SECONDS,
-				'display' => __( 'WP Cron Control Revisited internal job - every minute', 'wp-cron-control-revisited' ),
-			),
-			'wpccrij_ten_minutes' => array(
-				'interval' => 10 * MINUTE_IN_SECONDS,
-				'display' => __( 'WP Cron Control Revisited internal job - every 10 minutes', 'wp-cron-control-revisited' ),
-			),
-		);
 
 		// Prime lock cache if not present
 		wp_cache_add( $this->cache_key_lock, 0 );
@@ -156,26 +128,6 @@ class Main {
 			'permission_callback' => array( $this, 'check_secret' ),
 			'show_in_index'       => false,
 		) );
-	}
-
-	/**
-	 * Include custom schedules used for internal jobs
-	 */
-	public function register_internal_events_schedules( $schedules ) {
-		return array_merge( $schedules, $this->internal_jobs_schedules );
-	}
-
-	/**
-	 * Schedule internal jobs
-	 */
-	public function schedule_internal_events() {
-		$when = strtotime( sprintf( '+%d seconds', $this->job_queue_window_in_seconds ) );
-
-		foreach ( $this->internal_jobs as $job_args ) {
-			if ( ! wp_next_scheduled( $job_args['action'] ) ) {
-				wp_schedule_event( $when, $job_args['schedule'], $job_args['action'] );
-			}
-		}
 	}
 
 	/**
@@ -228,7 +180,7 @@ class Main {
 			foreach ( $timestamp_events as $action => $action_instances ) {
 				foreach ( $action_instances as $instance => $instance_args ) {
 					// There are some jobs we never care to run
-					if ( $this->is_blocked_event( $action ) ) {
+					if ( is_blocked_event( $action ) ) {
 						wp_unschedule_event( $timestamp, $action, $instance_args['args'] );
 						continue;
 					}
@@ -243,7 +195,7 @@ class Main {
 					);
 
 					// Queue internal events separately to avoid them being blocked
-					if ( $this->is_internal_event( $action ) ) {
+					if ( is_internal_event( $action ) ) {
 						$internal_events[] = $event;
 					} else {
 						$current_events[] = $event;
@@ -300,7 +252,7 @@ class Main {
 		$time_start = microtime( true );
 
 		// Limit how many events are processed concurrently
-		if ( ! $this->is_internal_event( $event['action'] ) && ! $this->check_lock() ) {
+		if ( ! is_internal_event( $event['action'] ) && ! $this->check_lock() ) {
 			return new WP_Error( 'no-free-threads', __( 'No resources available to run this job.', 'wp-cron-control-revisited' ) );
 		}
 
@@ -322,7 +274,7 @@ class Main {
 		do_action_ref_array( $event['action'], $event['args'] );
 
 		// Free process for the next event
-		if ( ! $this->is_internal_event( $event['action'] ) ) {
+		if ( ! is_internal_event( $event['action'] ) ) {
 			$this->free_lock();
 		}
 
@@ -376,22 +328,6 @@ class Main {
 	}
 
 	/**
-	 * Events that are always run, regardless of how many jobs are queued
-	 */
-	private function is_internal_event( $action ) {
-		return in_array( $action, wp_list_pluck( $this->internal_jobs, 'action' ) );
-	}
-
-	/**
-	 * Allow specific events to be blocked perpetually
-	 */
-	private function is_blocked_event( $action ) {
-		$blocked_hooks = array();
-
-		return in_array( $action, $blocked_hooks );
-	}
-
-	/**
 	 * Set a lock and limit how many concurrent jobs are permitted
 	 */
 	private function check_lock() {
@@ -430,51 +366,6 @@ class Main {
 		wp_cache_set( $this->cache_key_lock_timestamp, time() );
 
 		return true;
-	}
-
-	/**
-	 * Published scheduled posts that miss their schedule
-	 */
-	public function force_publish_missed_schedules() {
-		global $wpdb;
-
-		$missed_posts = $wpdb->get_col( $wpdb->prepare( "SELECT ID FROM {$wpdb->posts} WHERE post_status = 'future' AND post_date <= %s LIMIT 100;", current_time( 'mysql', false ) ) );
-
-		if ( ! empty( $missed_posts ) ) {
-			foreach ( $missed_posts as $missed_post ) {
-				check_and_publish_future_post( $missed_post );
-
-				do_action( 'wpccr_published_post_that_missed_schedule', $missed_post );
-			}
-		}
-	}
-
-	/**
-	 * Ensure scheduled posts have a corresponding cron job to publish them
-	 */
-	public function confirm_scheduled_posts() {
-		global $wpdb;
-
-		$future_posts = $wpdb->get_results( $wpdb->prepare( "SELECT ID, post_date FROM {$wpdb->posts} WHERE post_status = 'future' AND post_date > %s LIMIT 100;", current_time( 'mysql', false ) ) );
-
-		if ( ! empty( $future_posts ) ) {
-			foreach ( $future_posts as $future_post ) {
-				$future_post->ID = absint( $future_post->ID );
-				$gmt_time        = strtotime( get_gmt_from_date( $future_post->post_date ) . ' GMT' );
-				$timestamp       = wp_next_scheduled( 'publish_future_post', array( $future_post->ID ) );
-
-				if ( false === $timestamp ) {
-					wp_schedule_single_event( $gmt_time, 'publish_future_post', array( $future_post->ID ) );
-
-					do_action( 'wpccr_publish_scheduled', $future_post->ID );
-				} elseif ( (int) $timestamp !== $gmt_time ) {
-					wp_clear_scheduled_hook( 'publish_future_post', array( (int) $future_post->ID ) );
-					wp_schedule_single_event( $gmt_time, 'publish_future_post', array( $future_post->ID ) );
-
-					do_action( 'wpccr_publish_rescheduled', $future_post->ID );
-				}
-			}
-		}
 	}
 }
